requirements.txt
flask
flask-cors
google-generativeai
Pillow
python-dotenv
gunicorn
